Updating main matching file
Now includes a module skill so that the matching scheme may be customized per module. 

There are both while and for loops in the code. In order to run module A and B skills at once in the code we would need to nest the current loops within another which makes the code error prone. Instead we add another global variable which is the module skill which can be read in from Shiny. Then, simply call the code twice to produce output for each module.